requirements.txt
Pandas
Flask
RateMyProfessorAPI